Refactored actions logic into it's own component
