+++ Scripts/Actions/DebugMenuItemNode.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace DUCK.DebugMenu
+namespace DUCK.DebugMenu.Actions
 {
 	public enum NodeType
 	{
